requirements.txt
imagezmq
imutils
pyzmq
numpy
opencv-python
PyQt5
psycopg2-binary
face_recognition
torch
torchvision